webserver refactor / clean
