--- src/main/java/net/pms/webserver/HttpServletHelper.java
@@ -1,55 +1,0 @@
-/*
- * Universal Media Server, for streaming any media to DLNA
- * compatible renderers based on the http://www.ps3mediaserver.org.
- * Copyright (C) 2012 UMS developers.
- *
- * This program is a free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; version 2
- * of the License only.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- */
-package net.pms.webserver;
-
-import java.util.AbstractMap;
-import java.util.ArrayList;
-import java.util.Enumeration;
-import java.util.List;
-import java.util.Map;
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
-
-public class HttpServletHelper {
-
-	public static List<Map.Entry<String, String>> getHttpHeaders(HttpServletRequest request) {
-		List<Map.Entry<String, String>> headers = new ArrayList<>();
-		for (Enumeration<String> names = request.getHeaderNames(); names.hasMoreElements();) {
-			String name = (String) names.nextElement();
-			for (Enumeration<String> values = request.getHeaders(name); values.hasMoreElements();) {
-				headers.add(new AbstractMap.SimpleEntry<>(name, (String) values.nextElement()));
-			}
-		}
-		return headers;
-	}
-
-	public static String getCookie(String name, HttpServletRequest request) {
-		Cookie[] cookies = request.getCookies();
-		if (cookies != null) {
-			for (Cookie cookie : cookies) {
-				if (cookie.getName().equals(name)) {
-					return cookie.getValue();
-				}
-			}
-		}
-		return null;
-	}
-
-}
