+++ src/main/java/net/pms/webserver/handlers/FileHandler.java
@@ -17,74 +17,75 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package net.pms.webserver.servlets;
+package net.pms.webserver.handlers;
 
 import com.samskivert.mustache.MustacheException;
+import com.sun.net.httpserver.Headers;
+import com.sun.net.httpserver.HttpExchange;
+import com.sun.net.httpserver.HttpHandler;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.net.CookieHandler;
 import java.net.CookieManager;
 import java.net.HttpCookie;
 import java.net.HttpURLConnection;
-import java.net.URI;
 import java.net.URL;
 import java.net.URLConnection;
 import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.List;
-import java.util.stream.Collectors;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import net.pms.PMS;
 import net.pms.network.HTTPResource;
-import net.pms.webserver.RemoteUtil;
-import net.pms.webserver.WebServerServlets;
+import net.pms.webserver.WebServerUtil;
+import net.pms.webserver.WebServerHttpServer;
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class RemoteFileServlet extends WebServerServlet {
-	private static final Logger LOGGER = LoggerFactory.getLogger(RemoteFileServlet.class);
+public class FileHandler implements HttpHandler {
 
-	public RemoteFileServlet(WebServerServlets parent) {
-		super(parent);
+	private static final Logger LOGGER = LoggerFactory.getLogger(ThumbHandler.class);
+
+	private final WebServerHttpServer parent;
+
+	public FileHandler(WebServerHttpServer parent) {
+		this.parent = parent;
 	}
 
 	@Override
-	protected void service(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+	public void handle(HttpExchange t) throws IOException {
 		try {
-			URI uri = URI.create(request.getRequestURI());
-			LOGGER.debug("Handling web interface file request \"{}\"", uri);
+			LOGGER.debug("Handling web interface file request \"{}\"", t.getRequestURI());
 
-			String path = uri.getPath();
-			String responseString = null;
+			String path = t.getRequestURI().getPath();
+			String response = null;
 			String mime = null;
 			int status = 200;
 
 			if (path.contains("crossdomain.xml")) {
-				responseString = "<?xml version=\"1.0\"?>" +
-					"<!-- http://www.bitsontherun.com/crossdomain.xml -->" +
-					"<cross-domain-policy>" +
-					"<allow-access-from domain=\"*\" />" +
-					"</cross-domain-policy>";
+				response = "<?xml version=\"1.0\"?>" +
+						"<!-- http://www.bitsontherun.com/crossdomain.xml -->" +
+						"<cross-domain-policy>" +
+						"<allow-access-from domain=\"*\" />" +
+						"</cross-domain-policy>";
 				mime = "text/xml";
 
 			} else if (path.startsWith("/files/log/")) {
 				String filename = path.substring(11);
 				if (filename.equals("info")) {
 					String log = PMS.get().getFrame().getLog();
-					log = log.replace("\n", "<br>");
+					log = log.replaceAll("\n", "<br>");
 					String fullLink = "<br><a href=\"/files/log/full\">Full log</a><br><br>";
 					String x = fullLink + log;
 					if (StringUtils.isNotEmpty(log)) {
 						x = x + fullLink;
 					}
-					responseString = "<html><title>UMS LOG</title><body>" + x + "</body></html>";
+					response = "<html><title>UMS LOG</title><body>" + x + "</body></html>";
 				} else {
 					File file = parent.getResources().getFile(filename);
 					if (file != null) {
@@ -92,15 +93,15 @@ public class RemoteFileServlet extends WebServerServlet {
 						HashMap<String, Object> vars = new HashMap<>();
 						vars.put("title", filename);
 						vars.put("brush", filename.endsWith("debug.log") ? "debug_log" : filename.endsWith(".log") ? "log" : "conf");
-						vars.put("log", RemoteUtil.read(file).replace("<", "&lt;"));
-						responseString = parent.getResources().getTemplate("util/log.html").execute(vars);
+						vars.put("log", WebServerUtil.read(file).replace("<", "&lt;"));
+						response = parent.getResources().getTemplate("util/log.html").execute(vars);
 					} else {
 						status = 404;
 					}
 				}
 				mime = "text/html";
 			} else if (path.startsWith("/files/proxy")) {
-				String url = uri.getQuery();
+				String url = t.getRequestURI().getQuery();
 				if (url != null) {
 					url = url.substring(2);
 				}
@@ -112,21 +113,25 @@ public class RemoteFileServlet extends WebServerServlet {
 					CookieHandler.setDefault(cookieManager);
 				}
 
-				switch (request.getMethod()) {
+				switch (t.getRequestMethod()) {
 					case "POST":
+						ByteArrayOutputStream bytes = new ByteArrayOutputStream();
 						byte[] buf = new byte[4096];
 						int n;
-						String str = request.getReader().lines().collect(Collectors.joining(System.lineSeparator()));
+						while ((n = t.getRequestBody().read(buf)) > -1) {
+							bytes.write(buf, 0, n);
+						}
+						String str = bytes.toString("utf-8");
 						URLConnection conn = new URL(url).openConnection();
 						((HttpURLConnection) conn).setRequestMethod("POST");
-						conn.setRequestProperty("Content-type", request.getHeader("Content-type"));
+						conn.setRequestProperty("Content-type", t.getRequestHeaders().getFirst("Content-type"));
 						conn.setRequestProperty("Content-Length", String.valueOf(str.length()));
 						conn.setDoOutput(true);
 						OutputStreamWriter writer = new OutputStreamWriter(conn.getOutputStream());
 						writer.write(str);
 						writer.flush();
 						in = conn.getInputStream();
-						ByteArrayOutputStream bytes = new ByteArrayOutputStream();
+						bytes = new ByteArrayOutputStream();
 						while ((n = in.read(buf)) > -1) {
 							bytes.write(buf, 0, n);
 						}
@@ -151,13 +156,18 @@ public class RemoteFileServlet extends WebServerServlet {
 						}
 						break;
 				}
-				response.setContentType("text/plain");
-				response.addHeader("Access-Control-Allow-Origin", "*");
-				response.addHeader("Access-Control-Allow-Headers", "User-Agent");
-				response.addHeader("Access-Control-Allow-Headers", "Content-Type");
-				RemoteUtil.dumpDirect(in, response);
+				Headers hdr = t.getResponseHeaders();
+				hdr.add("Content-Type", "text/plain");
+				hdr.add("Access-Control-Allow-Origin", "*");
+				hdr.add("Access-Control-Allow-Headers", "User-Agent");
+				hdr.add("Access-Control-Allow-Headers", "Content-Type");
+				t.sendResponseHeaders(200, in.available());
+
+				OutputStream os = t.getResponseBody();
+				LOGGER.trace("input is {} output is {}", in, os);
+				WebServerUtil.dump(in, os);
 				return;
-			} else if (parent.getResources().write(path.substring(7), response)) {
+			} else if (parent.getResources().write(path.substring(7), t)) {
 				// The resource manager found and sent the file, all done.
 				return;
 
@@ -165,19 +175,19 @@ public class RemoteFileServlet extends WebServerServlet {
 				status = 404;
 			}
 
-			if (status == 404 && responseString == null) {
-				responseString = "<html><body>404 - File Not Found: " + path + "</body></html>";
+			if (status == 404 && response == null) {
+				response = "<html><body>404 - File Not Found: " + path + "</body></html>";
 				mime = "text/html";
 			}
-			RemoteUtil.respond(response, responseString, status, mime);
+
+			WebServerUtil.respond(t, response, status, mime);
 		} catch (IOException e) {
 			throw e;
 		} catch (MustacheException e) {
 			// Nothing should get here, this is just to avoid crashing the
 			// thread
-			LOGGER.error("Unexpected error in RemoteFileHandler.handle(): {}", e.getMessage());
+			LOGGER.error("Unexpected error in FileHandler.handle(): {}", e.getMessage());
 			LOGGER.trace("", e);
 		}
 	}
-
 }
