+++ src/main/java/net/pms/webserver/handlers/RawHandler.java
@@ -17,18 +17,17 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package net.pms.webserver.servlets;
+package net.pms.webserver.handlers;
 
+import com.sun.net.httpserver.Headers;
+import com.sun.net.httpserver.HttpExchange;
+import com.sun.net.httpserver.HttpHandler;
 import java.io.BufferedOutputStream;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.net.URI;
 import java.util.List;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import net.pms.PMS;
 import net.pms.configuration.WebRender;
 import net.pms.dlna.DLNAResource;
@@ -41,37 +40,40 @@ import net.pms.image.ImageInfo;
 import net.pms.image.ImagesUtil.ScaleType;
 import net.pms.io.OutputParams;
 import net.pms.io.ProcessWrapper;
-import net.pms.webserver.RemoteUtil;
-import net.pms.webserver.WebServerServlets;
+import net.pms.webserver.WebServerUtil;
+import net.pms.webserver.WebServerHttpServer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class RawServlet extends WebServerServlet {
-	private static final Logger LOGGER = LoggerFactory.getLogger(RawServlet.class);
+public class RawHandler implements HttpHandler {
+	private static final Logger LOGGER = LoggerFactory.getLogger(RawHandler.class);
 
-	public RawServlet(WebServerServlets parent) {
-		super(parent);
+	private final WebServerHttpServer parent;
+
+	public RawHandler(WebServerHttpServer parent) {
+		this.parent = parent;
 	}
 
 	@Override
-	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+	public void handle(HttpExchange t) throws IOException {
 		try {
-			URI uri = URI.create(request.getRequestURI());
-			LOGGER.debug("got a raw request " + uri);
-			RootFolder root = parent.getRoot(request, response);
-			if (root == null) {
-				LOGGER.debug("root not found");
-				response.sendError(401, "Unknown root");
-				return;
+			LOGGER.debug("got a raw request " + t.getRequestURI());
+			if (WebServerUtil.deny(t)) {
+				throw new IOException("Access denied");
 			}
-			String id = RemoteUtil.strip(RemoteUtil.getId("raw/", uri));
+
+			RootFolder root = parent.getRoot(WebServerUtil.userName(t), t);
+			if (root == null) {
+				throw new IOException("Unknown root");
+			}
+			String id;
+			id = WebServerUtil.strip(WebServerUtil.getId("raw/", t));
 			LOGGER.debug("raw id " + id);
 			List<DLNAResource> res = root.getDLNAResources(id, false, 0, 0, root.getDefaultRenderer());
 			if (res.size() != 1) {
 				// another error
 				LOGGER.debug("media unkonwn");
-				response.sendError(404, "Bad id");
-				return;
+				throw new IOException("Bad id");
 			}
 			DLNAResource dlna = res.get(0);
 			long len;
@@ -112,7 +114,7 @@ public class RawServlet extends WebServerServlet {
 			} else {
 				len = dlna.length();
 				dlna.setPlayer(null);
-				range = RemoteUtil.parseRange(request.getHeader("Range"), len);
+				range = WebServerUtil.parseRange(t.getRequestHeaders(), len);
 				in = dlna.getInputStream(range, root.getDefaultRenderer());
 				if (len == 0) {
 					// For web resources actual length may be unknown until we open the stream
@@ -121,28 +123,27 @@ public class RawServlet extends WebServerServlet {
 				mime = root.getDefaultRenderer().getMimeType(dlna);
 			}
 
+			Headers hdr = t.getResponseHeaders();
 			LOGGER.debug("Sending media \"{}\" with mime type \"{}\"", dlna, mime);
-			response.setContentType(mime);
-			response.addHeader("Accept-Ranges", "bytes");
-			response.addHeader("Server", PMS.get().getServerName());
-			response.addHeader("Connection", "keep-alive");
-			response.addHeader("Transfer-Encoding", "chunked");
+			hdr.add("Content-Type", mime);
+			hdr.add("Accept-Ranges", "bytes");
+			hdr.add("Server", PMS.get().getServerName());
+			hdr.add("Connection", "keep-alive");
+			hdr.add("Transfer-Encoding", "chunked");
 			if (in != null && in.available() != len) {
-				response.addHeader("Content-Range", "bytes " + range.getStart() + "-" + in.available() + "/" + len);
-				response.setContentLength(in.available());
-				response.setStatus(206);
+				hdr.add("Content-Range", "bytes " + range.getStart() + "-" + in.available() + "/" + len);
+				t.sendResponseHeaders(206, in.available());
 			} else {
-				response.setStatus(200);
+				t.sendResponseHeaders(200, 0);
 			}
-
-			OutputStream os = new BufferedOutputStream(response.getOutputStream(), 512 * 1024);
+			OutputStream os = new BufferedOutputStream(t.getResponseBody(), 512 * 1024);
 			LOGGER.debug("start raw dump");
-			RemoteUtil.dumpDirect(in, os);
+			WebServerUtil.dump(in, os);
 		} catch (IOException e) {
 			throw e;
 		} catch (InterruptedException e) {
 			// Nothing should get here, this is just to avoid crashing the thread
-			LOGGER.error("Unexpected error in RawServlet.doGet(): {}", e.getMessage());
+			LOGGER.error("Unexpected error in RawHandler.handle(): {}", e.getMessage());
 			LOGGER.trace("", e);
 		}
 	}
